Add User API, Protected Route, and Modify Upload Middleware

+++ react-query/backend/controllers/eventControllers.js
@@ -7,7 +7,7 @@ const fs = require("fs");
 
 // @desc    Create New Event
 // @route   POST /api/event
-// @access Public
+// @access  Private
 const createEvent = asyncHandler(async (req, res) => {
   const { title, description, eventDate, eventTime, eventLocation, eventImage } = req.body;
   if (!title || !eventDate || !eventTime || !eventLocation) {
@@ -20,6 +20,7 @@ const createEvent = asyncHandler(async (req, res) => {
     eventTime,
     eventLocation,
     eventImage,
+    userId: req?.user?.userId,
   });
 
   res.status(201).json({ data: events });
@@ -27,7 +28,7 @@ const createEvent = asyncHandler(async (req, res) => {
 
 // @desc    Upload Event Image
 // @route   POST /api/event/upload
-// @access Public
+// @access  Private
 const uploadEventImage = asyncHandler(async (req, res) => {
   try {
     await uploadFile(req, res);
@@ -72,7 +73,7 @@ const getAllEvents = asyncHandler(async (req, res) => {
 
 // @desc    Get Event by Id
 // @route   GET /api/event/:id
-// @access Public
+// @access  Public
 const getEventById = asyncHandler(async (req, res) => {
   const eventsId = req.params.id;
   if (!eventsId) {
@@ -89,7 +90,7 @@ const getEventById = asyncHandler(async (req, res) => {
 
 // @desc    Update Event
 // @route   PUT /api/event/:id
-// @access Public
+// @access  Private
 const updateEvent = asyncHandler(async (req, res) => {
   const eventsId = req.params.id;
   if (!eventsId) {
@@ -101,21 +102,25 @@ const updateEvent = asyncHandler(async (req, res) => {
     return res.status(404).json({ message: "Event Not Found" });
   }
 
-  const { title, description, eventDate, eventTime, eventLocation, eventImage } = req.body;
-  existingEvent.title = title || existingEvent.title;
-  existingEvent.description = description || existingEvent.description;
-  existingEvent.eventDate = eventDate || existingEvent.eventDate;
-  existingEvent.eventTime = eventTime || existingEvent.eventTime;
-  existingEvent.eventLocation = eventLocation || existingEvent.eventLocation;
-  existingEvent.eventImage = eventImage || existingEvent.eventImage;
+  if (parseInt(existingEvent.userId) == parseInt(req?.user?.userId)) {
+    const { title, description, eventDate, eventTime, eventLocation, eventImage } = req.body;
+    existingEvent.title = title || existingEvent.title;
+    existingEvent.description = description || existingEvent.description;
+    existingEvent.eventDate = eventDate || existingEvent.eventDate;
+    existingEvent.eventTime = eventTime || existingEvent.eventTime;
+    existingEvent.eventLocation = eventLocation || existingEvent.eventLocation;
+    existingEvent.eventImage = eventImage || existingEvent.eventImage;
 
-  await existingEvent.save();
-  res.status(200).json({ message: "Event Updated" });
+    await existingEvent.save();
+    res.status(200).json({ message: "Event Updated" });
+  }
+
+  return res.status(401).json({ message: "UnAuthorized" });
 });
 
 // @desc    Delete Event
 // @route   DELETE /api/event/:id
-// @access Public
+// @access  Private
 const deleteEvent = asyncHandler(async (req, res) => {
   const eventsId = req.params.id;
   if (!eventsId) {
@@ -127,15 +132,19 @@ const deleteEvent = asyncHandler(async (req, res) => {
     return res.status(404).json({ message: "Event Not Found" });
   }
 
-  const deletedEventImagePath = path.join(__dirname, "..", "public/uploads/", existingEvent.eventImage);
-  await fs.unlink(deletedEventImagePath, (err) => {
-    if (err) {
-      return res.status(500).json({ message: err });
-    }
-  });
+  if (parseInt(existingEvent.userId) == parseInt(req?.user?.userId)) {
+    const deletedEventImagePath = path.join(__dirname, "..", "public/uploads/", existingEvent.eventImage);
+    await fs.unlink(deletedEventImagePath, (err) => {
+      if (err) {
+        return res.status(500).json({ message: err });
+      }
+    });
 
-  await existingEvent.destroy();
-  res.status(200).json({ message: "Event Deleted" });
+    await existingEvent.destroy();
+    res.status(200).json({ message: "Event Deleted" });
+  }
+
+  return res.status(401).json({ message: "UnAuthorized" });
 });
 
 module.exports = {
